01.02.03.03,  01.02.04.03.01
 01.02.03.03
Created Summary View for MFI Setup Checklist, Separated steps into 8
phases, added pages up to Phase 4 Book 2.

01.02.04.03.01
Created Reviewer application module to allow the reviewer to add open
items / review points using a “Review swatch which is overlayed above
the application and doesnt effect the normal operation of the app. This
module is designed to be removed in the client version of the app. A
word doc has been added to the main folder outlining the code for
implementing the review module into any application.