00.01 Added resolution column to Open Items Page
Added resolution column to Open Items Page, also added a View/Add
button to be inserted by the refreshopenitems function for each open
item. In addition I added a viewResolutions() function that is called
when each button is clicked. The button passed the id of the open item
to the viewResolution function and sends a JSON request to a script on
the server. Currently the server script that will return all
resolutions for this item isnt built yet so the json request is
directed at:
http://www.findmydriver.com/ubmopenitemsallitems.php?callback=?

I will continue this at a later date.