Updated Model Dashboard Page
Added back buttons where needed.
Added overflow css to side panels too allow in depended scrolling of
page when necessary.
Changed the Resolution button on the open items page to now say
“Resolve” instead of “Add”
Fixed the list view on the open items resolution popup to allow the
both comments and resolutions to be added to the list.
Added User Preferences page with ability to add cell phone number to
receive notifications from the app. The field doesn’t currently update
the user preference in the database, this functionality needs to be
added. Edited the Navigation panel so it doesn’t feel so cluttered.
Added the ubmsuite_SelectBusinessModels page to the navigation panels.
Added conceptual definition form to the model settings page.
Added the identification setup form to the model settings page.
Deleted the identification setup page.
Deleted the conceptual definition page.