Added link to Edit Model Page, As per mrgoodkat9 request, open item # 30
Added a button that links to open the edit model page on all model
creation suite pages.

The user may experience unexpected behavior when pushing the back
button on the edit model page because if the user got there from a page
inside the MCS checklist the back button will take them back the the
page they were on instead of the My Models page.