Added UBM Product Creator.
UBM Product creator uses a wysiwyg editor to allow users to create html
products with an experience nearly comparable to what they would get in
Microsoft word or Open Office.
NOTE:
The Save function doesn’t currently work and will need to be modified
to allow cross domain transfer of information and storage of the
contents of each product in the api.universalbusinessmodel.com database.