Modified the Create model and create position functions and php scripts to store the hierarchical relationship in the closure table.
Added UBMS functionality to create model.
Added UBMS functionality to create position.

NOTE: the get my model positions has not been updated to use the
closure table so when you create a new position it doesn’t show up in
the application but if you see the popup it has been created it in the
closure table.