Closure Phase 1
Changes:
Updated the create model request to insert an instance in the hierarchy
closure table. updated app notifications to indicate as such.
Modified the getMyModels and getSharedModels to reference the UUID of
the model instead of the of the id.
Modified the setActiveModel to set the id of the activeModelUUID
instead of the activeModelId.

Replaced all instances of window.activeModelId with activeModelUUID

Updated getMyModelsOrgChart() function

Updated the projected financial statement function and iframe to use
activeModelUUID instead of activeModelId