Major changes and updates to the hierarchical object tree as well as the org chart builder.
Modified the height parameter for nodes on the org chart to auto resize
node when text begins to overflow.
Increased the width parameter for nodes on the org chart.

Removed several deprecated functions.

Updated the Hierarchical object tree to build itself dynamically and
correctly.

Added the universal unique position reference to each node on the org
chart.

Removed unnecessary forms on various popups shown on the org chart page.
Improved the look and feel of the same popups.

Updated getChecklistItems to allow submittal of prepared by records to
the DB.

Modified the setActiveTaskId in mcsChecklist to resolve conflicting
name with the Model Creation Suite Checklist.