Changed user sign in.
Changed user sign in so when enter is pressed and no password is
present it will take you to the password field instead of the username
field. Username field is no longer erased when you click sign in and
focus is automatically set to the password field. If the password is
wrong it now empties and focuses on the password field.