Finished functionality of task not complete
Made so when a reviewer clicks the red minus they get a popup asking if
the task is complete or incomplete. If the task is incomplete it
requires an action required before submitting to the database. Prepared
by script now checks to see if the prepared by record exists and
updates it instead of creating a new one.